% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanardAbsurd-package.R
\docType{package}
\name{CanardAbsurd-package}
\alias{CanardAbsurd}
\alias{CanardAbsurd-package}
\title{Durable R workflows with DuckDB and Quack}
\description{
A package for hosting a workflow database and running thin R workers. Its
checkpoint model is inspired by Absurd (\url{https://github.com/earendil-works/absurd}).
See \code{\link[=ca_serve]{ca_serve()}} for hosting, \code{\link[=ca_connect]{ca_connect()}} for remote access, \code{\link[=ca_work]{ca_work()}} for
workers, and \code{\link[=ca_step]{ca_step()}} for replay semantics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rgenomicsetl.github.io/CanardAbsurd/}
  \item \url{https://github.com/RGenomicsETL/CanardAbsurd}
  \item Report bugs at \url{https://github.com/RGenomicsETL/CanardAbsurd/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}
}

}
\keyword{internal}
